[domain] move RemoteKey into module
