+++ src/main/scala/net/kemitix/s3thorp/Config.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp
 
 import java.io.File
 
-import net.kemitix.s3thorp.domain.Bucket
+import net.kemitix.s3thorp.domain.{Bucket, RemoteKey}
 
 final case class Config(
   bucket: Bucket = Bucket(""),
