+++ src/main/scala/net/kemitix/s3thorp/KeyModified.scala
@@ -1,5 +1,7 @@
 package net.kemitix.s3thorp
 
+import net.kemitix.s3thorp.domain.RemoteKey
+
 final case class KeyModified(key: RemoteKey,
                              modified: LastModified) {
