+++ src/test/scala/net/kemitix/s3thorp/DummyS3Client.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp
 
 import cats.effect.IO
 import net.kemitix.s3thorp.awssdk.{S3Client, S3ObjectsData, UploadProgressListener}
-import net.kemitix.s3thorp.domain.Bucket
+import net.kemitix.s3thorp.domain.{Bucket, RemoteKey}
 
 trait DummyS3Client extends S3Client {
