+++ src/test/scala/net/kemitix/s3thorp/S3ActionSuite.scala
@@ -1,5 +1,7 @@
 package net.kemitix.s3thorp
 
+import net.kemitix.s3thorp.domain.RemoteKey
+
 class S3ActionSuite extends UnitTest {
 
   describe("Ordering of types") {
